Feature/password hash2

+++ src/main/java/com/example/demo/controller/UserPageController.java
@@ -2,6 +2,7 @@ package com.example.demo.controller;
 
 import com.example.demo.entity.User;
 import com.example.demo.service.UserService;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -20,7 +21,13 @@ public class UserPageController {
 
     @GetMapping()//moved it up
     public String adminUserPage(@RequestParam(required = false) String keyword,
+                                HttpSession session,
                                 Model model) {
+
+        User user = (User) session.getAttribute("user");
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
 
         List<User> users;
 
@@ -32,13 +39,21 @@ public class UserPageController {
 
         model.addAttribute("users", users);
         model.addAttribute("keyword", keyword);
+        model.addAttribute("username", user.getName());
+        model.addAttribute("userRole", user.getRole().name());
 
         return "adminUserPage";
     }
 
     @PostMapping("/update-role/{id}")
     public String updateUserRole(@PathVariable Long id,
-                                 @RequestParam String role) {
+                                 @RequestParam String role,
+                                 HttpSession session) {
+
+        User user = (User) session.getAttribute("user");
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
 
         userService.updateUserRole(
                 id,
@@ -49,7 +64,13 @@ public class UserPageController {
     }
 
     @PostMapping("/delete/{id}")
-    public String deleteUser(@PathVariable Long id) {
+    public String deleteUser(@PathVariable Long id,
+                             HttpSession session) {
+
+        User user = (User) session.getAttribute("user");
+        if (user == null || user.getRole() != User.Role.ADMIN) {
+            return "redirect:/loginpage";
+        }
 
         userService.deleteUser(id);
